refactor to set language after call done
